Button Bug
Found a bug where if I set the setTimeout() function to a longer time then the user will be able to click the button multiple times.

Then once the setTimeout() function is done there will be a delay and all those delayed setTimeout() functions will trigger.

Fix: Added a boolean check to see if this event listener was triggered and if so break out of the function.